Remove read module, move it back into server.rs
