+++ crates/fluke/src/h2/mod.rs
@@ -8,5 +8,4 @@ pub(crate) mod parse;
 
 mod body;
 mod encode;
-mod read;
 mod types;
